Lookup host_name and hostgroup_name in service templates generating relationships
This is mandatory for assigning services to hosts and hostgroups in the
first place. Icinga 2 cannot later detect whether this attribute is
hidden in a template or not, but requires the direct service->host and
service->hostgroup relation.

Therefore the attributes need to be looked up in the parent objects
which could cost some additional cpu time due to recursive lookup but
still paves the way for the notifications and contact* attribute search
too.